Add kas::event::handler module for Handler trait
